+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- An error would occur if only one item was in a collection
+
 ## [1.0.5] - 2020-05-24
 
 ### Changed